refactor: Handle notification channel creation during notification request
