+++ app/src/main/java/campus/tech/kakao/map/MapFirebaseMessagingService.kt
@@ -1,5 +1,6 @@
 package campus.tech.kakao.map
 
+import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
@@ -19,6 +20,7 @@ class MapFirebaseMessagingService : FirebaseMessagingService() {
         }
 
         if (remoteMessage.notification != null) {
+            createNotificationChannel()
             sendNotification()
         }
     }
@@ -27,12 +29,7 @@ class MapFirebaseMessagingService : FirebaseMessagingService() {
         val intent = Intent(this, SplashActivity::class.java).apply {
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
         }
-        val pendingIntent = PendingIntent.getActivity(
-            this,
-            0,
-            intent,
-            PendingIntent.FLAG_IMMUTABLE
-        )
+        val pendingIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_IMMUTABLE)
         val builder = NotificationCompat.Builder(this, CHANNEL_ID)
             .setSmallIcon(R.drawable.ic_launcher_foreground)
             .setContentTitle("[중요] 포그라운드 알림")
@@ -49,10 +46,20 @@ class MapFirebaseMessagingService : FirebaseMessagingService() {
         notificationManager.notify(NOTIFICATION_ID, builder.build())
     }
 
+    private fun createNotificationChannel() {
+        val name = getString(R.string.channel_name)
+        val descriptionText = getString(R.string.channel_description)
+        val importance = NotificationManager.IMPORTANCE_DEFAULT
+        val channel = NotificationChannel(CHANNEL_ID, name, importance)
+        channel.description = descriptionText
+        val notificationManager: NotificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.createNotificationChannel(channel)
+    }
+
+
     companion object {
-        private const val NOTIFICATION_ID = 222222
-        private const val CHANNEL_ID = "main_default_channel"
-        private const val CHANNEL_NAME = "main channelName"
+        private const val NOTIFICATION_ID = 1
+        private const val CHANNEL_ID = "notification_default_channel"
     }
 
 }
