+++ app/src/main/res/values/strings.xml
@@ -3,5 +3,6 @@
     <string name="search_placeholder">검색어를 입력해 주세요.</string>
     <string name="search_no_data">검색 결과가 없습니다.</string>
     <string name="BASE_URL">https://dapi.kakao.com</string>
-    <string name="channel_description">notification</string>
+    <string name="channel_name">notification</string>
+    <string name="channel_description">defalut_notification</string>
 </resources>
